Fix User

+++ model/modelUser.php
@@ -2,66 +2,88 @@
 include('../../utility/mysql_db.php');
 class modelUser extends mysql_db
 {
-    public function bacauakpb()
+    public function bacaunitsatker()
     {
-        $query = "select * from satker
-                    where kd_uapb is not null and
-                          kd_uappbe1 is not null and
-                          kd_uappbw is not null and
-                          kd_uakpb is not null and
-                          kd_uapkpb is not null and
-                          jk is not null";
+        $query = "select kodesektor, namasatker from satker
+                        where kodesektor is not null and
+                              kodesatker is null and
+                              kodeunit is null and
+                              gudang is null 
+                        order by kodesektor asc";
         $result = $this->query($query);
-        echo '<option value="">-- Pilih Kode Satker--</option>';
+        echo '<option value="">-- Pilih Kode Unit --</option>';
         while ($row = $this->fetch_array($result))
         {
-            // $replace = str_replace(array('.', ','), '' , $row['kode']);
-            echo '<option value="'.$row['kode'].'">'.$row['kode'].' '.$row['nm_satker']."</option>";
+            if (!empty($row)) {
+                echo '<optgroup label="'.$row['kodesektor'].' '.$row['namasatker'].'">';
+                $query2 = "select kodesektor, kodesatker, namasatker from satker
+                        where kodesektor ='$row[kodesektor]' and
+                              kodesatker is not null and
+                              kodeunit is null and
+                              gudang is null 
+                        order by kodesektor asc";
+                $result2 = $this->query($query2);
+                while ($row2 = $this->fetch_array($result2))
+                {
+                    if (!empty($row2)) {
+                        echo '<optgroup label="&nbsp;&nbsp;'.$row2['kodesektor'].'.'.$row2['kodesatker'].' '.$row2['namasatker'].'">';
+                        $query3 = "select kodesektor, kodesatker, kodeunit, namasatker from satker
+                                where kodesektor ='$row2[kodesektor]' and
+                                      kodesatker ='$row2[kodesatker]' and
+                                      kodeunit is not null and
+                                      gudang is null 
+                                order by kodesektor asc";
+                        $result3 = $this->query($query3);
+                        while ($row3 = $this->fetch_array($result3))
+                        {
+                            echo '<option value="'.$row3['kodesektor'].'.'.$row3['kodesatker'].'.'.$row3['kodeunit'].'">&nbsp;&nbsp;'.$row3['kodesektor'].'.'.$row3['kodesatker'].'.'.$row3['kodeunit'].' '.$row3['namasatker']."</option>";
+                        }
+                        echo '</optgroup>';
+                    }
+                }
+                echo '</optgroup>';
+            }
         }
     }
     public function bacadata($data)
     {
-        $sql1="select kd_uapb from satker
+        $sql1="select kodesektor from satker
                     where kode = '$data'";
-        $sql2="select kd_uappbe1 from satker
+        $sql2="select kodesatker from satker
                     where kode = '$data'";
-        $sql3="select kd_uappbw from satker
-                    where kode = '$data'";
-        $sql4="select jk from satker
+        $sql3="select kodeunit from satker
                     where kode = '$data'";
         $result1 = $this->query($sql1);
         $result2 = $this->query($sql2);
         $result3 = $this->query($sql3);
+        $sektor = mysqli_fetch_assoc($result1);
+        $satker = mysqli_fetch_assoc($result2);
+        $unit = mysqli_fetch_assoc($result3);
+        $kodesektor = $sektor["kodesektor"];
+        $kodesatker = $satker["kodesatker"];
+        $kodeunit = $unit["kodeunit"];
+        $sql4="select namasatker from satker
+                    where kodesektor = '$kodesektor' and
+                        kodesatker is null and
+                        kodeunit is null and
+                        gudang is null";
+        $sql5="select namasatker from satker
+                    where kodesektor = '$kodesektor' and
+                        kodesatker = '$kodesatker' and
+                        kodeunit is null and
+                        gudang is null";
+        $sql6="select namasatker from satker
+                    where kodesektor = '$kodesektor' and
+                        kodesatker = '$kodesatker' and
+                        kodeunit = '$kodeunit' and
+                        gudang is null";
         $result4 = $this->query($sql4);
-        $uapb = mysqli_fetch_assoc($result1);
-        $uappbe = mysqli_fetch_assoc($result2);
-        $uappbw = mysqli_fetch_assoc($result3);
-        $kdjk = mysqli_fetch_assoc($result4);
-        $kodeuapb = $uapb["kd_uapb"];
-        $kodeuappbe = $uappbe["kd_uappbe1"];
-        $kodeuappbw = $uappbw["kd_uappbw"];
-        $sql5="select nm_satker from satker
-                    where kd_uapb = '$kodeuapb' and
-                        kd_uappbe1 is null and
-                        kd_uappbw is null and
-                        kd_uakpb is null";
-        $sql6="select nm_satker from satker
-                    where kd_uapb = '$kodeuapb' and
-                        kd_uappbe1 = '$kodeuappbe' and
-                        kd_uappbw is null and
-                        kd_uakpb is null";
-        $sql7="select nm_satker from satker
-                    where kd_uapb = '$kodeuapb' and
-                        kd_uappbe1 = '$kodeuappbe' and
-                        kd_uappbw = '$kodeuappbw' and
-                        kd_uakpb is null";
         $result5 = $this->query($sql5);
         $result6 = $this->query($sql6);
-        $result7 = $this->query($sql7);
-        $uruapb = mysqli_fetch_assoc($result5);
-        $uruappbe = mysqli_fetch_assoc($result6);
-        $uruappbw = mysqli_fetch_assoc($result7);
-        echo json_encode(array("kduapb"=>$uapb["kd_uapb"],"kduappbe"=>$uappbe["kd_uappbe1"],"kduappbw"=>$uappbw["kd_uappbw"],"kdjk"=>$kdjk["jk"],"uruapb"=>$uruapb["nm_satker"],"uruappbe"=>$uruappbe["nm_satker"],"uruappbw"=>$uruappbw["nm_satker"]));
+        $ursektor = mysqli_fetch_assoc($result4);
+        $ursatker = mysqli_fetch_assoc($result5);
+        $urunit = mysqli_fetch_assoc($result6);
+        echo json_encode(array("kdsektor"=>$sektor["kodesektor"],"kdsatker"=>$sektor["kodesektor"].$satker["kodesatker"],"kdunit"=>$sektor["kodesektor"].$satker["kodesatker"].$unit["kodeunit"],"ursektor"=>$ursektor["namasatker"],"ursatker"=>$ursatker["namasatker"],"urunit"=>$urunit["namasatker"]));
     }
 	public function tambahuser($data)
 	{
@@ -70,12 +92,7 @@ class modelUser extends mysql_db
 		$user_pass = $data['user_pass'];
         $user_email = $data['user_email'];
         $kd_satker= $data['kd_lokasi'];
-        echo "kd_satker : ".$kd_satker;
-        $query_satker = "select nm_satker from satker where kode='$kd_satker'";
-        $result_satker = $this->query($query_satker);
-        $data_satker = mysqli_fetch_assoc($result_satker);
-        $nm_satker = $data_satker["nm_satker"];
-
+        $nm_satker= $data['nm_satker'];
 		$kd_lokasi = str_replace(array('.'), '' , $data['kd_lokasi']);
 		$user_level = 2;
 		$query = "Insert into user
@@ -108,36 +125,9 @@ class modelUser extends mysql_db
     public function hapususer($data)
     {
 
-        $user_name = $data['user_name'];
-        $user_pass = $data['user_pass'];
-        $user_email = $data['user_email'];
-        $user_level = $data['user_level'];
-        $query = "delete from user
-                    where user_id='$user_id'";
+        $query = "delete from user where user_id='$data'";
         $result = $this->query($query);
         return $result;
-    }
-
-    public function bacaUser()
-    {
-        $query = "select * from user";
-        $result = $this->query($query);
-        echo '<option value="">-- Pilih User Id --</option>';
-        while ($row = $this->fetch_array($result))
-        {
-            // echo '<option value="'.$row['user_id'].'">'.$row['user_name'].' '.$row['user_pass'].' '.$row['user_level']"</option>";
-        }   
-    }
-    public function bacatable($data)
-    {
-        $query = "select * from user
-                    where user_id = '$data'";
-        $result = $this->query($query);
-        while ($row = $this->fetch_assoc($result))
-        {
-            // $rows[] = [$row['user_id'],$row["user_name"],$row["user_pass"],$row["user_level"]];
-        }
-        echo json_encode($rows);
     }
 }
 ?>
